[#20] Create detail ui
