+++ lib/crawler_web/templates/home/_keyword_list.html.heex
@@ -21,7 +21,7 @@
             <%= keyword.id %>
           </td>
           <td class="w-50">
-            <%= keyword.name %>
+            <%= link(keyword.name, to: Routes.keyword_path(@conn, :show, keyword)) %>
           </td>
           <td class={get_column_status(keyword.status)}>
             <%= get_status(keyword.status) %>
