+++ lib/crawler_web/views/keyword_view.ex
@@ -1,3 +1,5 @@
 defmodule CrawlerWeb.KeywordView do
   use CrawlerWeb, :view
+
+  import CrawlerWeb.ViewHelpers
 end
