+++ lib/crawler_web/views/view_helpers.ex
@@ -1,0 +1,25 @@
+defmodule CrawlerWeb.ViewHelpers do
+  import CrawlerWeb.Gettext
+
+  def is_completed(status), do: status == :completed
+
+  def get_status(value) do
+    case value do
+      :new -> gettext("New")
+      :in_progress -> gettext("Crawling")
+      :failed -> gettext("Failed")
+      :completed -> gettext("Completed")
+      _ -> gettext("Undefined")
+    end
+  end
+
+  def get_column_status(value) do
+    case value do
+      :new -> "text-info"
+      :in_progress -> "text-warning"
+      :failed -> "text-danger"
+      :completed -> "text-success"
+      _ -> "text-muted"
+    end
+  end
+end
