+++ priv/gettext/default.pot
@@ -88,7 +88,7 @@ msgid "Keyword"
 msgstr ""
 
 #: lib/crawler_web/controllers/api/v1/keyword_controller.ex:72
-#: lib/crawler_web/controllers/keyword_controller.ex:61
+#: lib/crawler_web/controllers/keyword_controller.ex:66
 #, elixir-autogen, elixir-format
 msgid "One or more keywords are invalid! Allowed keyword length is %{min}-%{max}"
 msgstr ""
@@ -120,13 +120,13 @@ msgid "The file doesn't look like CSV"
 msgstr ""
 
 #: lib/crawler_web/controllers/api/v1/keyword_controller.ex:58
-#: lib/crawler_web/controllers/keyword_controller.ex:47
+#: lib/crawler_web/controllers/keyword_controller.ex:52
 #, elixir-autogen, elixir-format
 msgid "The file is empty"
 msgstr ""
 
 #: lib/crawler_web/controllers/api/v1/keyword_controller.ex:64
-#: lib/crawler_web/controllers/keyword_controller.ex:53
+#: lib/crawler_web/controllers/keyword_controller.ex:58
 #, elixir-autogen, elixir-format
 msgid "The file is too big, allowed size is up to %{limit} keywords"
 msgstr ""
@@ -141,7 +141,7 @@ msgstr ""
 msgid "Uploaded file is not a valid CSV file"
 msgstr ""
 
-#: lib/crawler_web/controllers/keyword_controller.ex:33
+#: lib/crawler_web/controllers/keyword_controller.ex:38
 #, elixir-autogen, elixir-format
 msgid "%{num_keyword} keywords were uploaded!"
 msgstr ""
@@ -161,27 +161,107 @@ msgstr ""
 msgid "Unauthorized"
 msgstr ""
 
-#: lib/crawler_web/views/home_view.ex:19
+#: lib/crawler_web/views/view_helpers.ex:11
 #, elixir-autogen, elixir-format
 msgid "Completed"
 msgstr ""
 
-#: lib/crawler_web/views/home_view.ex:17
+#: lib/crawler_web/views/view_helpers.ex:9
 #, elixir-autogen, elixir-format
 msgid "Crawling"
 msgstr ""
 
-#: lib/crawler_web/views/home_view.ex:18
+#: lib/crawler_web/views/view_helpers.ex:10
 #, elixir-autogen, elixir-format
 msgid "Failed"
 msgstr ""
 
-#: lib/crawler_web/views/home_view.ex:16
+#: lib/crawler_web/views/view_helpers.ex:8
 #, elixir-autogen, elixir-format
 msgid "New"
 msgstr ""
 
-#: lib/crawler_web/views/home_view.ex:20
+#: lib/crawler_web/views/view_helpers.ex:12
 #, elixir-autogen, elixir-format
 msgid "Undefined"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:27
+#, elixir-autogen, elixir-format
+msgid "Created At: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:19
+#, elixir-autogen, elixir-format
+msgid "ID: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:14
+#, elixir-autogen, elixir-format
+msgid "Keyword Info"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:52
+#, elixir-autogen, elixir-format
+msgid "Non-Ad Count: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:93
+#, elixir-autogen, elixir-format
+msgid "Non-Ad URLs"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:41
+#, elixir-autogen, elixir-format
+msgid "Result Report"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:75
+#, elixir-autogen, elixir-format
+msgid "Search Result HTML"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:65
+#, elixir-autogen, elixir-format
+msgid "Searching"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:23
+#, elixir-autogen, elixir-format
+msgid "Status: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:129
+#, elixir-autogen, elixir-format
+msgid "There are no Non-Ad URLs"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:113
+#, elixir-autogen, elixir-format
+msgid "There are no Top-Ad URLs"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:46
+#, elixir-autogen, elixir-format
+msgid "Top Ad Count: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:84
+#, elixir-autogen, elixir-format
+msgid "Top Ad URLs"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:49
+#, elixir-autogen, elixir-format
+msgid "Total Ad Count: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:55
+#, elixir-autogen, elixir-format
+msgid "Total: %{value}"
+msgstr ""
+
+#: lib/crawler_web/templates/keyword/show.html.heex:30
+#, elixir-autogen, elixir-format
+msgid "Updated At: %{value}"
 msgstr ""
